rename taskQueue module to kebap-case
